==> source/cases/_R/insurance/ibli/2-modis.rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# MODIS data pre-processing

In this chapter we download, pre-process and compute NDVI from MODIS data. 
These procedures are discussed in more detail [here](https://rspatial.org/terra/modis/index.html). 

## Area of interest

First we get the area of interest. Normally you would get this from a file (for example, a shapefile). See `?terra::vect` for reading such files.

```{r aoi, message=FALSE}
library(agrodata)
aoi <- data_ibli("marsabit") 
aoi
```

As you can see we have 58 polygons, each representing a "sub-location". Let's plot the polygons on top of a map of Kenya to see where they are.

```{r ken, message=FALSE}
library(geodata)
ken <- gadm("KEN", level=1, path=".")
plot(ken, border="gray")
plot(ken[ken$NAME_1 == "Marsabit", ], col="beige", border="gray", add=TRUE)
lines(aoi, col="red")
```

So we are in Northe Kenya. Also note that some polygons (sub-locations) are much smaller than others.

## Get MODIS files

Here we show how to download and pre-process the data, but only for 1 month, to limit the download and processing needed. After that we use pre-processed data.

Set up the download directory

```{r downloaddir}
datadir <- "modis"
dir.create(datadir, FALSE)
# datadir
```

In this example we download data from 2010 only. 

```{r dates}
start <- "2010-01-01" 
end <- "2010-01-07"
```

To download the data you need to provide the username and password for your (free) EOSDIS account. If you do not have an account, you can sign up [here](https://urs.earthdata.nasa.gov/users/new). Here we use passwords that are stored in a file that I read in below (sorry, we cannot show you the values).

```{r pwd}
up <- readRDS("../pwds.rds")
up <- up[up$service == "EOSDIS", ]
```

Now we are ready to get the data

```{r moddown, message=FALSE}
library(luna)
fmod <- getModis(product="MOD09A1", start, end, aoi, download=TRUE, 
			path=datadir, username=up$user, password=up$pwd)
fmod
```

We only have four files here, each representing 8 days -- just to illustrate the processing.

A single MODIS tile for extends well over the study area. We first crop the data to correspond to the study area, and then we remove bad pixels (clouds and shadow) and then .

## Quality

To remove bad quality pixels from the selected files, we follow the same method described [here](https://rspatial.org/terra/modis/4-quality.html). We specify a matrix in which each row has start and end of the quality assessment (QA) bits considered, and the values to be rejected (in the image).

```{r qcMat}
from <- c(1,3,11,12) 
to   <- c(2,6,11,14) 
reject <- c("10,11", "1100,1101,1110,1111", "1", "000,110,111")
qmat <- cbind(from, to, reject)
```

We need to project the case study boundary to the sinusoidal coordinate reference system that the MODIS data are in.

```{r prj}
prj <- "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs "
pols <- project(aoi, prj)
```

## Monthly NDVI computation 

We now compute NDVI for each month. First, produce a quality mask, mask out bad quality pixels, crop to the area of interest, and then compute NDVI. 

We use a loop over all files (`fmod`).

```{r Preprocess}
for(i in seq_along(fmod)) {
  # Load a MODIS file
  r <- rast(fmod[i])

  # Generate quality mask (layer 12)
  quality_mask <- modis_mask(r[[12]], 16, qmat)

  #Select only red and NIR
  r <- r[[1:2]]
  names(r) <- c("red", "NIR")

  # remove low quality pixels
  r <- mask(r, quality_mask)

  # remove areas outside the AOI rectangle
  r <- crop(r, pols)

  # ensure that all values are between 0 and 1 
  r <- clamp(r, 0, 1) 

  # Compute NDVI
  ndvi <- (r$NIR - r$red)/(r$NIR + r$red)

  filename <- gsub(".hdf$", "_ndvi.tif", fmod[i])

  writeRaster(ndvi, filename=filename, overwrite=TRUE)
}
```
